+++ beltradar/models/__init__.py
@@ -1,3 +1,3 @@
 # flake8: noqa
+from .beltradar import *
 from .general import *
-from .owner import *