+++ CHANGELOG.md
@@ -11,6 +11,14 @@ Section Order:
 ### Removed
 -->
 
+### Added
+
+- Display estimated completion time for belt mining
+- Display remaining volume and belt size information
+- Show mining speed in m³/s
+- Permission System to view and manage Survey Sessions
+- Prepared Task Function
+
 ## [0.0.1] - 2024-08-xx
 
 ### Added